Revert "Revert "monet: colors: Switch to single-precision floating-point math""
This reverts commit e5f5792ee7650ec59693a37ce684772f6f9334c8.

On second thought, let's ship this for 5.1.1 and reevaluate it in
colorkt.

+++ app/src/main/java/dev/kdrag0n/android12ext/monet/colors/Cam16.kt
@@ -3,14 +3,14 @@ package dev.kdrag0n.android12ext.monet.colors
 import dev.kdrag0n.android12ext.monet.colors.google.Cam
 
 data class Cam16(
-    override val L: Double,
-    override val C: Double,
-    override val h: Double,
+    override val L: Float,
+    override val C: Float,
+    override val h: Float,
 ) : Lch {
     override fun toLinearSrgb() = toSrgb().toLinearSrgb()
 
     fun toSrgb(): Srgb {
-        val cam = Cam.fromJch(L.toFloat(), C.toFloat(), h.toFloat())
+        val cam = Cam.fromJch(L, C, h)
         return Srgb(cam.viewedInSrgb())
     }
 
@@ -18,9 +18,9 @@ data class Cam16(
         fun Srgb.toCam16(): Cam16 {
             val cam = Cam.fromInt(quantize8())
             return Cam16(
-                L = cam.j.toDouble(),
-                C = cam.chroma.toDouble(),
-                h = cam.hue.toDouble(),
+                L = cam.j,
+                C = cam.chroma,
+                h = cam.hue,
             )
         }
     }
